requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
pydub==0.25.1
python-dotenv==1.0.1
